Fix other incorrect narrowing for format strings
Note that zd is for Py_ssize_t, and `NPY_INTP_FMT` is for npy_intp. In all conceivable situations these are the same type, but we don't today check that at compile time, so for now we maintain the distinction.

Found by searching for `Format\([^;]+\(int\)`.

+++ numpy/core/src/multiarray/conversion_utils.c
@@ -667,8 +667,8 @@ PyArray_ConvertClipmodeSequence(PyObject *object, NPY_CLIPMODE *modes, int n)
     if (object && (PyTuple_Check(object) || PyList_Check(object))) {
         if (PySequence_Size(object) != n) {
             PyErr_Format(PyExc_ValueError,
-                    "list of clipmodes has wrong length (%d instead of %d)",
-                    (int)PySequence_Size(object), n);
+                    "list of clipmodes has wrong length (%zd instead of %d)",
+                    PySequence_Size(object), n);
             return NPY_FAIL;
         }
